% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drisk.R
\name{drscore}
\alias{drscore}
\title{Disclosure Risk}
\usage{
drscore(
  Sample,
  Protected,
  delta = 0.05,
  neighbourhood = 1,
  kdistinct = 5,
  ldeniable = kdistinct,
  neigh_type = "constant",
  numeric.vars = NULL,
  outlier.par = list(centre = median, scale = var, thresh = 0.01)
)
}
\arguments{
\item{direction}{a numeric value to set the direction of the color scale  if direction is < 0 switch the direction}

\item{...}{other arguments to \code{\link[ggplot2]{discrete_scale}}}
}
\description{
\code{drscore} Disclosure Risk
}
\details{
Default colors are generated with branded palette. A categorical variable is discrete in the sense that it has clear boundaries from one ategory to another. On the other hand, a continuous scale does not have such clear separation. Hence, different colors can be used for different categories and one color with different gradient across the scale can be used for continuous variable.
}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
