% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drisk.R
\name{drscore}
\alias{drscore}
\title{Disclosure Risk}
\usage{
drscore(Sample,Protected,
delta = 0.05,
neighbourhood = 1,
kdistinct = 5,
ldeniable = kdistinct,
neigh_type = "constant",
numeric.vars = NULL,
outlier.par = list(centre = median,
                   scale = var,
                   thresh = 0.01))
}
\arguments{
\item{Sample}{A sample of the original dataset. A data frame or data frame extension (e.g. a tibble), matrix or an array,}

\item{Protected}{A sample of the protected dataset. A data frame or data frame extension (e.g. a tibble), matrix or an array.}

\item{delta}{amount of difference}

\item{neighbourhood}{Possible 'neighbourhood' types
1 = Mahalanobis (Based on Mahalanobis Distance)
2 = DSTAR   (Based on Density Based Distance)
3 = StdEuclid (Based on Standardised (by std dev) Euclidean Distance)
4 = RelEuclid (Relative Euclidean Distance sum_k ((Xk-Yk)/Xk)^2)}

\item{kdistinct}{number of data point in the neighbourhood to be classified as protected. k distinct threshold if integer then probability threshold is k/SS (SS = sample size)}

\item{ldeniable}{methods of define disclosure. l undeniable threshold if integer then probability threshold is l/SS (SS = sample size)}

\item{neigh_type}{methods of define length/area of neighbourhood. Possible 'neigh_type' types
constant = fixed threshold on distance
prob = Nearest Neighbour Probability Neighbourhood used (Worst Case Scenario 1)
estprob = Nearest Neighbour Probability Neighbourhood used based on protected density (Worst Case Scenario 2)}

\item{numeric.vars}{#Continuous Variables in the dataset.}

\item{outlier.par}{A list(centre, scale, thresh). Define the method of determining outlier. Parameters to adjust how MV outliers are determined. Default is that lie 99 percent (based on Chi-Square n-1 dist) away from median after scale by variance.}
}
\value{
An object of the type as `DRisk`. Summary of the disclosure risk in terms of distinctness, similarities and undeniable parameters.
}
\description{
\code{drscore} Disclosure Risk
}
\details{
Statistical disclosure control methods for micro-level continuous data are both varied and complex. This method can be applied to any pair of original and protected data-sets despite a difference in dimensionality and without assuming any particular joint probability structure between the original and protected data.
}
\examples{
# ##################
# ##all continuous###################

library(svMisc)
library(dress)
library(sdcMicro)

CASC_sample <- CASCrefmicrodata[,c(2,3,4,6)]
CASC_protected <- addNoise(CASC_sample,noise = 100)$xm #Additive Noise protected

DRisk_NN <- drscore(
  Sample = CASC_sample, #Original Sample
  Protected = CASC_protected,
  delta = 0.05,
  kdistinct = 0.05, #k distinct threshold if integer then
  # probability threshold is k/SS (SS = sample size)
  ldeniable = 5, # l undeniable threshold if integer then
  # probability threshold is l/SS (SS = sample size)
  neighbourhood = 1,
  #Possible 'neighbourhood' types
  # 1 = Mahalanobis (Based on Mahalanobis Distance)
  # 2 = DSTAR   (Based on Density Based Distance)
  # 3 = StdEuclid (Based on Standardised (by std dev) Euclidean Distance)
  # 4 = RelEuclid (Relative Euclidean Distance sum_k ((Xk-Yk)/Xk)^2)
  neigh_type = 'prob',
  #Possible 'neigh_type' types
  #constant = fixed threshold on distance
  #prob = Nearest Neighbour Probability Neighbourhood used
  #estprob = = Nearest Neighbour Probability Neighbourhood used based on protected density
  numeric.vars = 1:4, #Which Variables are continuous?
  outlier.par = list(centre = median,
                     scale = var,
                     thresh = 0.01)
  #Parameters to adjust how MV outliers are determined.
  #Default is that lie 99\% (based on Chi-Square n-1 dist) away from median after scale by variance.
)


}
\seealso{
\code{\link[dress]{update}}
}
