% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update}
\usage{
update(DRisk, ...)
}
\arguments{
\item{DRisk}{results from drscore. Drisk object.}

\item{...}{other arguments to \code{\link[dress]{drscore}}}
}
\description{
\code{update} Update
}
\examples{
########################
## mixed dataset
########################
library(svMisc)
library(dress)
nn <- drscore(Sample = wage4, Protected = wage4_protected, numeric.vars = c("age","wage"))
#xd <- update(nn,delta = 0.06)





}
\seealso{
\code{\link[ggplot2]{discrete_scale}}
}
