% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wage4.R
\docType{data}
\name{wage4}
\alias{wage4}
\title{wage4
`r class(wage4)`}
\format{
A data frame with `r nrow(wage4)` rows and `r ncol(wage4)` variables:
\describe{
  \item{entity}{country}
  \item{code}{3 letter country code}
  \item{year}{year from 1961 to 2018}
  \item{tonnes_per_hectare}{produce of crop in tonnes per hectare}
  ...
}
}
\source{
\url{'https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-09-01/key_crop_yields.csv'}
}
\usage{
wage4
}
\description{
More details about the data
}
\section{Data Contents}{

The data was collected from... during...
}

\examples{
head(wage4)
}
\keyword{datasets}
