% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wage4_protected.R
\docType{data}
\name{wage4_protected}
\alias{wage4_protected}
\title{wage4_protected}
\format{
A data frame with 3000 observations and 4 variables:
\describe{
  \item{age}{Age of worker}
  \item{education}{A factor with levels 1. < HS Grad 2. HS Grad 3. Some College 4. College Grad and 5. Advanced Degree indicating education level}
  \item{jobclass}{A factor with levels 1. Industrial and 2. Information indicating type of job}
  \item{wage}{Workers raw wage}
  ...
}
}
\source{
Data was manually assembled by Steve Miller, of Inquidia Consulting (formerly Open BI). From the March 2011 Supplement to Current Population Survey data.
\url{'https://www.re3data.org/repository/r3d100011860'}
}
\usage{
wage4_protected
}
\description{
Wage and other data protected with some disclosure control method for original the Mid-Atlantic region wage data.
}
\examples{
head(wage4_protected)
}
\references{
James, G., Witten, D., Hastie, T., and Tibshirani, R. (2013) An Introduction to Statistical Learning with applications in R, https://www.statlearning.com, Springer-Verlag, New York
}
\seealso{
\code{\link[ISLR]{Wage}} \code{\link[dress]{wage4}}
}
\keyword{datasets}
